Update game scene and Dockerfile, add keyboard event handling
This commit does several changes to improve the application. Updated components are removed from various nodes in the game scene. They seem to be redundant and unnecessary, making the scene cleaner. Docker file was updated to use a newer Debian image '11-slim' because it offers better security and support. Lastly, a 'useEffect' hook was used to prevent default actions for the 'space' keyboard event in the 'GameButton.tsx'. This will improve user experience, as the spacebar had unwanted side actions while playing the game.