Adjust Resume and RaydevsGame components for better UI interaction
Code edits were made to the Resume and RaydevsGame components to make them more responsive and interactive. The minHeight style property was moved to conditionally apply only when the game is not playing. This prevents the unused screen space from appearing when the game is active. Also, added conditional rendering for the Summary and Segment components depending on the 'isPlaying' state. In RaydevsGame component, 'position' style value changed from 'absolute' to 'relative' to control the game component's layout better. In the SegmentDescription.tsx file, each segment description item is given a unique key for better React element handling.